# backend/docs/api-docs.json
{"openapi":"3.1.0","info":{"title":"Food Guard API","description":"Food Guard backend API documentation","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Food","description":"사용자 식품 관리 API"},{"name":"Auth","description":"회원가입, 로그인, 닉네임 중복 확인 API"},{"name":"ChatMessage","description":"채팅 메시지 API"},{"name":"Comment","description":"게시글 댓글 API"},{"name":"Group","description":"그룹 관리 API"},{"name":"ChatRoom","description":"채팅방 API"},{"name":"Post","description":"게시글 API"},{"name":"User","description":"사용자 API"},{"name":"Donation","description":"나눔 게시글 API"}],"paths":{"/api/v1/posts/{id}":{"get":{"tags":["Post"],"summary":"게시글 단건 조회","description":"게시글 ID로 상세 정보를 조회합니다.","operationId":"getPost","parameters":[{"name":"id","in":"path","description":"게시글 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}}}},"put":{"tags":["Post"],"summary":"게시글 수정","description":"게시글 작성자가 게시글을 수정합니다.","operationId":"editPost","parameters":[{"name":"id","in":"path","description":"게시글 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostEditRequest"}}},"required":true},"responses":{"200":{"description":"수정 성공"}}},"delete":{"tags":["Post"],"summary":"게시글 삭제","description":"게시글 작성자가 게시글을 삭제합니다.","operationId":"deletePost","parameters":[{"name":"id","in":"path","description":"게시글 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"삭제 성공"}}}},"/api/v1/groups/{id}":{"get":{"tags":["Group"],"summary":"그룹 단건 조회","description":"그룹 ID로 상세 정보를 조회합니다.","operationId":"getGroup","parameters":[{"name":"id","in":"path","description":"그룹 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupResponse"}}}}}},"put":{"tags":["Group"],"summary":"그룹 수정","description":"그룹 관리자가 그룹 정보를 수정합니다.","operationId":"editGroup","parameters":[{"name":"id","in":"path","description":"그룹 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupEditRequest"}}},"required":true},"responses":{"200":{"description":"수정 성공"}}},"delete":{"tags":["Group"],"summary":"그룹 삭제","description":"그룹 관리자가 그룹을 삭제합니다.","operationId":"deleteGroup","parameters":[{"name":"id","in":"path","description":"그룹 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"삭제 성공"}}}},"/api/v1/foods/{id}":{"get":{"tags":["Food"],"summary":"식품 단건 조회","description":"식품 ID로 로그인한 사용자의 식품 상세 정보를 조회합니다.","operationId":"getFood","parameters":[{"name":"id","in":"path","description":"식품 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FoodResponse"}}}}}},"put":{"tags":["Food"],"summary":"식품 수정","description":"식품 ID로 로그인한 사용자의 식품 정보를 수정합니다.","operationId":"editFood","parameters":[{"name":"id","in":"path","description":"식품 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FoodEditRequest"}}},"required":true},"responses":{"200":{"description":"수정 성공"}}},"delete":{"tags":["Food"],"summary":"식품 삭제","description":"식품 ID로 로그인한 사용자의 식품을 삭제합니다.","operationId":"deleteFood","parameters":[{"name":"id","in":"path","description":"식품 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"삭제 성공"}}}},"/api/v1/donations/{id}":{"get":{"tags":["Donation"],"summary":"나눔 단건 조회","description":"나눔 ID로 상세 정보를 조회합니다.","operationId":"getDonation","parameters":[{"name":"id","in":"path","description":"나눔 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DonationResponse"}}}}}},"put":{"tags":["Donation"],"summary":"나눔 수정","description":"나눔 작성자가 나눔 게시글 정보를 수정합니다.","operationId":"editDonation","parameters":[{"name":"id","in":"path","description":"나눔 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DonationEditRequest"}}},"required":true},"responses":{"200":{"description":"수정 성공"}}},"delete":{"tags":["Donation"],"summary":"나눔 삭제","description":"나눔 작성자가 나눔 게시글을 삭제합니다.","operationId":"deleteDonation","parameters":[{"name":"id","in":"path","description":"나눔 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"삭제 성공"}}}},"/api/v1/comments/{id}":{"put":{"tags":["Comment"],"summary":"댓글 수정","description":"댓글 작성자가 댓글 내용을 수정합니다.","operationId":"editComment","parameters":[{"name":"id","in":"path","description":"댓글 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentEditRequest"}}},"required":true},"responses":{"200":{"description":"수정 성공"}}},"delete":{"tags":["Comment"],"summary":"댓글 삭제","description":"댓글 작성자가 댓글을 삭제합니다.","operationId":"deleteComment","parameters":[{"name":"id","in":"path","description":"댓글 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"삭제 성공"}}}},"/api/v1/posts":{"get":{"tags":["Post"],"summary":"게시글 목록 조회","description":"전체 게시글 목록을 조회합니다.","operationId":"getPosts","responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PostResponse"}}}}}}},"post":{"tags":["Post"],"summary":"게시글 작성","description":"로그인한 사용자가 게시글을 작성합니다.","operationId":"createPost","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostCreateRequest"}}},"required":true},"responses":{"200":{"description":"작성 성공"}}}},"/api/v1/posts/{postId}/comments":{"get":{"tags":["Comment"],"summary":"댓글 목록 조회","description":"게시글 ID로 댓글과 대댓글 목록을 조회합니다.","operationId":"getComments","parameters":[{"name":"postId","in":"path","description":"게시글 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponse"}}}}}}},"post":{"tags":["Comment"],"summary":"댓글 작성","description":"게시글에 댓글 또는 대댓글을 작성합니다.","operationId":"createComment","parameters":[{"name":"postId","in":"path","description":"게시글 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentCreateRequest"}}},"required":true},"responses":{"200":{"description":"작성 성공"}}}},"/api/v1/groups":{"post":{"tags":["Group"],"summary":"그룹 생성","description":"로그인한 사용자가 새 그룹을 생성합니다.","operationId":"createGroup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreateRequest"}}},"required":true},"responses":{"200":{"description":"생성 성공"}}}},"/api/v1/groups/{id}/exit":{"post":{"tags":["Group"],"summary":"그룹 탈퇴","description":"로그인한 사용자가 그룹에서 탈퇴합니다.","operationId":"groupExit","parameters":[{"name":"id","in":"path","description":"그룹 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"탈퇴 성공"}}}},"/api/v1/foods":{"get":{"tags":["Food"],"summary":"내 식품 목록 조회","description":"로그인한 사용자가 등록한 식품 목록을 조회합니다.","operationId":"getFoods","responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FoodResponse"}}}}}}},"post":{"tags":["Food"],"summary":"식품 등록","description":"로그인한 사용자의 식품을 등록합니다.","operationId":"createFood","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FoodCreateRequest"}}},"required":true},"responses":{"200":{"description":"등록 성공"}}}},"/api/v1/donations":{"get":{"tags":["Donation"],"summary":"나눔 목록 조회","description":"전체 나눔 게시글 목록을 조회합니다.","operationId":"getDonations","responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DonationResponse"}}}}}}},"post":{"tags":["Donation"],"summary":"나눔 등록","description":"로그인한 사용자가 보유 식품으로 나눔 게시글을 등록합니다.","operationId":"createDonation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DonationCreateRequest"}}},"required":true},"responses":{"201":{"description":"등록 성공"}}}},"/api/v1/donations/{donationId}/chatrooms":{"post":{"tags":["ChatRoom"],"summary":"채팅방 시작","description":"나눔 게시글에 대한 채팅방을 시작하거나 기존 채팅방을 조회합니다.","operationId":"startChatRoom","parameters":[{"name":"donationId","in":"path","description":"나눔 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"시작 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoomResponse"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Auth"],"summary":"회원가입","description":"아이디, 비밀번호, 닉네임으로 새 사용자를 생성합니다.","operationId":"signUp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}},"required":true},"responses":{"200":{"description":"회원가입 성공","content":{"application/json":{"schema":{"type":"string"}}}}},"security":[]}},"/api/v1/auth/login":{"post":{"tags":["Auth"],"summary":"로그인","description":"아이디와 비밀번호로 로그인하고 JWT access token을 발급받습니다.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"로그인 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"security":[]}},"/api/v1/auth/check/nickname":{"post":{"tags":["Auth"],"summary":"닉네임 중복 확인","description":"입력한 닉네임을 사용할 수 있는지 확인합니다.","operationId":"checkNickname","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckNicknameRequest"}}},"required":true},"responses":{"200":{"description":"사용 가능 여부 반환","content":{"application/json":{"schema":{"type":"boolean"}}}}},"security":[]}},"/api/v1/chatrooms/{id}/close":{"patch":{"tags":["ChatRoom"],"summary":"채팅방 종료","description":"채팅방을 종료 상태로 변경합니다.","operationId":"closeChatRoom","parameters":[{"name":"id","in":"path","description":"채팅방 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"종료 성공"}}}},"/api/v1/users/me":{"get":{"tags":["User"],"summary":"내 정보 조회","description":"로그인한 사용자의 정보를 조회합니다.","operationId":"findUserById","responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"delete":{"tags":["User"],"summary":"회원 탈퇴","description":"로그인한 사용자를 삭제 처리합니다.","operationId":"deleteUser","responses":{"200":{"description":"탈퇴 성공"}}}},"/api/v1/users/groups":{"get":{"tags":["User"],"summary":"내 그룹 목록 조회","description":"로그인한 사용자가 속한 그룹 목록을 조회합니다.","operationId":"findUserGroupsById","responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserGroupResponse"}}}}}}}},"/api/v1/groups/me":{"get":{"tags":["Group"],"summary":"내 그룹 목록 조회","description":"로그인한 사용자가 속한 그룹 목록을 조회합니다.","operationId":"getGroups","responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupResponse"}}}}}}}},"/api/v1/chatrooms":{"get":{"tags":["ChatRoom"],"summary":"내 채팅방 목록 조회","description":"로그인한 사용자의 채팅방 목록을 조회합니다.","operationId":"findMyChatRooms","responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatRoomResponse"}}}}}}}},"/api/v1/chatrooms/{id}":{"get":{"tags":["ChatRoom"],"summary":"내 채팅방 단건 조회","description":"채팅방 ID로 로그인한 사용자의 채팅방 상세 정보를 조회합니다.","operationId":"findMyChatRoom","parameters":[{"name":"id","in":"path","description":"채팅방 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRoomResponse"}}}}}}},"/api/v1/chatrooms/{chatRoomId}/messages":{"get":{"tags":["ChatMessage"],"summary":"채팅 메시지 목록 조회","description":"채팅방 ID로 메시지 목록을 조회합니다.","operationId":"findMessages","parameters":[{"name":"chatRoomId","in":"path","description":"채팅방 ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"조회 성공","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatMessageResponse"}}}}}}}}},"components":{"schemas":{"PostEditRequest":{"type":"object","description":"게시글 수정 요청","properties":{"title":{"type":"string","description":"게시글 제목","example":"남는 식재료 보관 팁"},"content":{"type":"string","description":"게시글 내용","example":"냉장고 식재료를 오래 보관하는 방법을 공유합니다."}}},"GroupEditRequest":{"type":"object","description":"그룹 수정 요청","properties":{"name":{"type":"string","description":"그룹명","example":"우리집"},"description":{"type":"string","description":"그룹 설명","example":"가족이 함께 식품을 관리하는 그룹"}}},"FoodEditRequest":{"type":"object","description":"식품 수정 요청","properties":{"name":{"type":"string","description":"식품명","example":"우유"},"type":{"type":"string","description":"식품 종류","example":"DAIRY"},"description":{"type":"string","description":"식품 설명","example":"개봉하지 않은 저지방 우유"},"expiryAt":{"type":"string","format":"date","description":"소비기한","example":"2026-06-30"},"groupId":{"type":"integer","format":"int64","description":"식품을 공유할 그룹 ID","example":1}}},"DonationEditRequest":{"type":"object","description":"나눔 수정 요청","properties":{"title":{"type":"string","description":"나눔 게시글 제목","example":"오늘까지인 샐러드 나눔합니다."},"content":{"type":"string","description":"나눔 게시글 내용","example":"회사 근처에서 직접 가져가실 수 있는 분께 나눔합니다."},"foodId":{"type":"integer","format":"int64","description":"나눔할 식품 ID","example":1}}},"CommentEditRequest":{"type":"object","description":"댓글 수정 요청","properties":{"content":{"type":"string","description":"수정할 댓글 내용","example":"수정된 댓글 내용입니다."}},"required":["content"]},"PostCreateRequest":{"type":"object","description":"게시글 작성 요청","properties":{"title":{"type":"string","description":"게시글 제목","example":"남는 식재료 보관 팁"},"content":{"type":"string","description":"게시글 내용","example":"냉장고 식재료를 오래 보관하는 방법을 공유합니다."}},"required":["content","title"]},"CommentCreateRequest":{"type":"object","description":"댓글 작성 요청","properties":{"content":{"type":"string","description":"댓글 내용","example":"좋은 정보 감사합니다."},"parentId":{"type":["integer","null"],"format":"int64","description":"부모 댓글 ID. 일반 댓글이면 null","example":1}},"required":["content"]},"GroupCreateRequest":{"type":"object","description":"그룹 생성 요청","properties":{"name":{"type":"string","description":"그룹명","example":"우리집"},"description":{"type":"string","description":"그룹 설명","example":"가족이 함께 식품을 관리하는 그룹"}},"required":["name"]},"FoodCreateRequest":{"type":"object","description":"식품 등록 요청","properties":{"name":{"type":"string","description":"식품명","example":"우유"},"type":{"type":"string","description":"식품 종류","example":"DAIRY"},"description":{"type":"string","description":"식품 설명","example":"개봉하지 않은 저지방 우유"},"expiryAt":{"type":"string","format":"date","description":"소비기한","example":"2026-06-30"},"groupId":{"type":"integer","format":"int64","description":"식품을 공유할 그룹 ID","example":1}},"required":["expiryAt","name","type"]},"DonationCreateRequest":{"type":"object","description":"나눔 등록 요청","properties":{"title":{"type":"string","description":"나눔 게시글 제목","example":"오늘까지인 샐러드 나눔합니다."},"content":{"type":"string","description":"나눔 게시글 내용","example":"회사 근처에서 직접 가져가실 수 있는 분께 나눔합니다."},"foodId":{"type":"integer","format":"int64","description":"나눔할 식품 ID","example":1}},"required":["content","foodId","title"]},"ChatRoomResponse":{"type":"object","description":"채팅방 응답","properties":{"id":{"type":"integer","format":"int64","description":"채팅방 ID","example":1},"donationId":{"type":"integer","format":"int64","description":"나눔 ID","example":1},"ownerId":{"type":"integer","format":"int64","description":"나눔 작성자 ID","example":1},"requesterId":{"type":"integer","format":"int64","description":"요청자 ID","example":2},"opponentId":{"type":"integer","format":"int64","description":"현재 사용자 기준 상대방 ID","example":2},"opponentNickname":{"type":"string","description":"현재 사용자 기준 상대방 닉네임","example":"나눔요청자"},"status":{"type":"string","description":"채팅방 상태","enum":["OPEN","CLOSED"],"example":"OPEN"},"lastMessage":{"type":["string","null"],"description":"마지막 메시지","example":"아직 나눔 가능할까요?"},"lastMessageAt":{"type":["string","null"],"format":"date-time","description":"마지막 메시지 작성일시","example":"2026-06-02T10:30:00"},"createdAt":{"type":"string","format":"date-time","description":"채팅방 생성일시","example":"2026-06-02T10:00:00"}}},"SignUpRequest":{"type":"object","description":"회원가입 요청","properties":{"username":{"type":"string","description":"로그인에 사용할 사용자 아이디","example":"foodguard01"},"password":{"type":"string","description":"로그인 비밀번호","example":"Password123!"},"nickname":{"type":"string","description":"서비스에 표시될 닉네임","example":"푸드가드"}},"required":["nickname","password","username"]},"LoginRequest":{"type":"object","description":"로그인 요청","properties":{"username":{"type":"string","description":"사용자 아이디","example":"foodguard01"},"password":{"type":"string","description":"비밀번호","example":"Password123!"}},"required":["password","username"]},"LoginResponse":{"type":"object","description":"로그인 응답","properties":{"accessToken":{"type":"string","description":"JWT access token","example":"eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIn0.signature"}}},"CheckNicknameRequest":{"type":"object","description":"닉네임 중복 확인 요청","properties":{"nickname":{"type":"string","description":"중복 확인할 닉네임","example":"푸드가드"}},"required":["nickname"]},"UserResponse":{"type":"object","description":"사용자 응답","properties":{"id":{"type":"integer","format":"int64","description":"사용자 ID","example":1},"username":{"type":"string","description":"사용자 아이디","example":"foodguard01"},"nickname":{"type":"string","description":"사용자 닉네임","example":"푸드가드"}}},"UserGroupResponse":{"type":"object","description":"사용자 그룹 응답","properties":{"id":{"type":"integer","format":"int64","description":"그룹 ID","example":1},"name":{"type":"string","description":"그룹명","example":"우리집"}}},"PostResponse":{"type":"object","description":"게시글 응답","properties":{"id":{"type":"integer","format":"int64","description":"게시글 ID","example":1},"title":{"type":"string","description":"게시글 제목","example":"남는 식재료 보관 팁"},"content":{"type":"string","description":"게시글 내용","example":"냉장고 식재료를 오래 보관하는 방법을 공유합니다."},"author":{"type":"string","description":"작성자 닉네임","example":"푸드가드"},"createdAt":{"type":"string","format":"date-time","description":"작성일시","example":"2026-06-02T10:30:00"}}},"CommentResponse":{"type":"object","description":"댓글 응답","properties":{"id":{"type":"integer","format":"int64","description":"댓글 ID","example":1},"content":{"type":"string","description":"댓글 내용","example":"좋은 정보 감사합니다."},"parentId":{"type":["integer","null"],"format":"int64","description":"부모 댓글 ID. 최상위 댓글이면 null","example":1},"status":{"type":"string","description":"댓글 상태","enum":["PUBLISHED","EDITED","DELETED"],"example":"PUBLISHED"},"updatedAt":{"type":"string","format":"date-time","description":"수정일시","example":"2026-06-02T10:30:00"},"children":{"type":"array","description":"대댓글 목록","items":{"$ref":"#/components/schemas/CommentResponse"}}}},"GroupResponse":{"type":"object","description":"그룹 응답","properties":{"id":{"type":"integer","format":"int64","description":"그룹 ID","example":1},"name":{"type":"string","description":"그룹명","example":"우리집"},"description":{"type":"string","description":"그룹 설명","example":"가족이 함께 식품을 관리하는 그룹"}}},"FoodResponse":{"type":"object","description":"식품 응답","properties":{"id":{"type":"integer","format":"int64","description":"식품 ID","example":1},"name":{"type":"string","description":"식품명","example":"우유"},"type":{"type":"string","description":"식품 종류","example":"DAIRY"},"description":{"type":"string","description":"식품 설명","example":"개봉하지 않은 저지방 우유"},"expiryAt":{"type":"string","format":"date","description":"소비기한","example":"2026-06-30"}}},"DonationResponse":{"type":"object","description":"나눔 응답","properties":{"postId":{"type":"integer","format":"int64","description":"게시글 ID","example":1},"title":{"type":"string","description":"나눔 게시글 제목","example":"오늘까지인 샐러드 나눔합니다."},"content":{"type":"string","description":"나눔 게시글 내용","example":"회사 근처에서 직접 가져가실 수 있는 분께 나눔합니다."},"author":{"type":"string","description":"작성자 닉네임","example":"푸드가드"},"createdAt":{"type":"string","format":"date-time","description":"작성일시","example":"2026-06-02T10:30:00"},"donationId":{"type":"integer","format":"int64","description":"나눔 ID","example":1},"foodId":{"type":"integer","format":"int64","description":"식품 ID","example":1},"foodName":{"type":"string","description":"식품명","example":"샐러드"},"status":{"type":"string","description":"나눔 상태","enum":["ONGOING","COMPLETED"],"example":"ONGOING"}}},"ChatMessageResponse":{"type":"object","description":"채팅 메시지 응답","properties":{"id":{"type":"integer","format":"int64","description":"메시지 ID","example":1},"chatRoomId":{"type":"integer","format":"int64","description":"채팅방 ID","example":1},"senderId":{"type":"integer","format":"int64","description":"발신자 ID","example":1},"senderNickname":{"type":"string","description":"발신자 닉네임","example":"푸드가드"},"content":{"type":"string","description":"메시지 내용","example":"아직 나눔 가능할까요?"},"type":{"type":"string","description":"메시지 타입","enum":["TEXT","SYSTEM"],"example":"TEXT"},"createdAt":{"type":"string","format":"date-time","description":"생성일시","example":"2026-06-02T10:30:00"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}