# backend/src/main/java/com/dna/fooo_guard/domain/auth/dto/LoginRequest.java
package com.dna.fooo_guard.domain.auth.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Schema(description = "로그인 요청")
@Getter
@Builder
@NoArgsConstructor(access = AccessLevel.PROTECTED)
@AllArgsConstructor(access = AccessLevel.PRIVATE)
public class LoginRequest {

    @Schema(description = "사용자 아이디", example = "foodguard01", requiredMode = Schema.RequiredMode.REQUIRED)
    private String username;

    @Schema(description = "비밀번호", example = "Password123!", requiredMode = Schema.RequiredMode.REQUIRED)
    private String password;
}